requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
appdirs==1.4.4
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cmake==3.27.5
comm==0.1.3
contourpy==1.1.0
croniter==1.3.15
cycler==0.11.0
dateutils==0.6.12
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.3.0
defusedxml==0.7.1
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.6.1
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.97.0
fastjsonschema==2.17.1
filelock==3.9.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.1.1
gitdb==4.0.10
gitignore-parser==0.1.3
GitPython==3.1.31
h11==0.14.0
huggingface-hub==0.15.1
idna==3.4
importlib-resources==5.12.0
inquirer==3.1.3
ipykernel==6.23.3
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonargparse==4.21.2
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lightning @ git+https://github.com/Lightning-AI/lightning.git@bdac61a11894264fcf2b40d174be99cb0c32e006
lightning-cloud==0.5.37
lightning-utilities==0.8.0
lit==17.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mpmath==1.2.1
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.0rc1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.1.78
ordered-set==4.1.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.8.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.7
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.1.0
pyre-extensions==0.0.29
pyrsistent==0.19.3
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.3
pytorch-triton==0.0.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
readchar==4.0.5
regex==2023.6.3
requests==2.28.1
rich==13.4.2
safetensors==0.3.1
scipy==1.11.1
segment-anything==1.0
sentry-sdk==1.26.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
sympy==1.11.1
timm @ git+https://github.com/huggingface/pytorch-image-models.git@9260cf517d6ba597b33ad1bbf7dbff9d0ccadd75
tinycss2==1.2.1
tokenizers==0.13.3
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
triton==2.0.0
typeshed-client==2.3.0
typing-inspect==0.9.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.22.0
wandb==0.15.4
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.0
websockets==11.0.3
xformers==0.0.20
yarl==1.9.2
